\name{om16}
\alias{om16}
\title{Orthogonal matrix for 16 runs}
\description{This function generates an orthogonal matrix in 16 runs with 6 columns with given numbers}
\usage{om16(v)}
\arguments{
	\item{v}{a vector containing numbers}	
	}
\value{
	\item{D}{(An orthogonal matrix of 16 runs and 6 columns}
  }
\examples{
om16(c(1,2,4,8,10,12,14,16))
}
\author{Baidya Nath Mandal<mandal.stat@gmail.com>}
\keyword{internal}